[pipelining] Rename to stage.py and schedules.py
[ghstack-poisoned]
