+++ test/distributed/pipelining/test_schedule.py
@@ -20,8 +20,8 @@ from torch.distributed.pipelining import (
     ScheduleInterleaved1F1B,
     ScheduleLoopedBFS,
 )
-from torch.distributed.pipelining.PipelineSchedule import _Action, _ComputationType
-from torch.distributed.pipelining.PipelineStage import _PipelineStageBase
+from torch.distributed.pipelining.schedules import _Action, _ComputationType
+from torch.distributed.pipelining.stage import _PipelineStageBase
 from torch.testing._internal.common_cuda import TEST_MULTIGPU
 from torch.testing._internal.common_distributed import (
     MultiProcContinousTest,
