+++ torch/distributed/pipelining/__init__.py
@@ -1,12 +1,12 @@
 # Copyright (c) Meta Platforms, Inc. and affiliates
 from ._IR import Pipe, pipe_split, pipeline, SplitPoint
-from .PipelineSchedule import (
+from .schedules import (
     Schedule1F1B,
     ScheduleGPipe,
     ScheduleInterleaved1F1B,
     ScheduleLoopedBFS,
 )
-from .PipelineStage import build_stage, PipelineStage
+from .stage import build_stage, PipelineStage
 
 __all__ = [
     "Pipe",
